==> Markdowns/03_Combine_Data.Rmd <==
---
title: "03_Combine_Data"
output:
  html_document
date: "2023-09-23"
---

```{r setup, include=FALSE}
source(paste0(getwd(), "/../Config/global_options.R"))
suppressMessages(source(paste0(getwd(), "/../Scripts/Functions.R")))
```


# Read in data

``` {r}
# Read in datasets from the saved RDS objects generated in 02_*_Data_Wrangling
lipidData <- readRDS(paste0(wd, "RDS/Data_Objects/cleaned_lipid_data.rds"))
proteinData <- readRDS(paste0(wd, "RDS/Data_Objects/cleaned_protein_data.rds"))

# Extract each of the metadata features 
lipidSampleNames <- lipidData$sampleNames

proteinSampleNames <- proteinData$sampleNames
proteinPeptideCounts <- proteinData$peptideCounts
proteinFragmentCounts <- proteinData$fragmentCounts


lipidData <- lipidData$df
proteinData <- proteinData$df
```

# Coerce dataframe rows into the same order

First, we convert the sample names from each dataframe into the same syntax:

``` {r}
# Remove first three characters, corresponds to sample number
lipidSampleNames <- substr(lipidSampleNames, 4, 100)
# Remove leading 0s from sample ID
lipidSampleNames <- gsub("(?<=_)0+", "", lipidSampleNames, perl = TRUE)

# All proteins have "_1" at end, remove this
proteinSampleNames <- substr(proteinSampleNames, 1, nchar(proteinSampleNames) - 2)
```

We can check the agreement between these two sets of sample names by sorting each vector of sample names and checking the pair-wise agreement between them. Notice the four sample names which do not agree. Three of them are the same and just in differing positions. What looks to be the issue is that in the lipid dataset there is a sample named `CAD_1265`. This same sample is named `CAD_265` in the protein dataset.

``` {r}
# Determine which samples do not agree across dataframes
discordantSampleNames <- sort(lipidSampleNames) != sort(proteinSampleNames)
sort(lipidSampleNames)[discordantSampleNames]
sort(proteinSampleNames)[discordantSampleNames]
```

Given that our assumption of a naming error is correct, we will adjust this name in the lipid dataset. Given the intersection of the two dataframe's sample names is of length 70, we now know that all the samples are in the same format and exist across both datasets:

``` {r}
# Adjust individual sample name
lipidSampleNames[lipidSampleNames=="CAD_1265"] <- "CAD_265"
# Check that the intersection is length 70
intersect(lipidSampleNames, proteinSampleNames) %>% length()
```

We can then make the dataframes' row orders agree:

``` {r}
# Set dataframes to have the same sample row order
lipidData <- lipidData[mixedorder(lipidSampleNames),]
proteinData <- proteinData[mixedorder(proteinSampleNames),]

sampleNames <- mixedsort(lipidSampleNames)
```

# Clean protein names

``` {r}
# Extract accession IDs
accessionIDs <- colnames(proteinData) %>% lapply(function(name) {
  strsplit(name, split="|", fixed=T)[[1]][2]
}) %>% unlist()

# Extract protein names and remove the _HUMAN at the end
colnames(proteinData) <- colnames(proteinData) %>% lapply(function(name) {
  strsplit(name, split="|", fixed=T)[[1]][3]
}) %>% unlist()
colnames(proteinData) <- gsub("_HUMAN", "", colnames(proteinData))
```




# Create numeric dataframes

While we already set these to numerics in sections `02.1` and `02.2`, they have have been converted back into numerics in above processing. Never hurts to ensure they are in the correct style:

``` {r}
# Coerce to numerics
lipidData <- lipidData %>% lapply(as.numeric) %>% as.data.table()
proteinData <- proteinData %>% lapply(as.numeric) %>% as.data.table()
```



# Impute missing values

Only lipid data has missing values. Impute with median value of each column

``` {r}
# Iterate through each variable. If there are NAs, replace these with the median of that variable
lipidData <- lipidData %>% lapply(function(v) {
  nas <- is.na(v)
  if (length(nas) > 0) {
    v[nas] <- median(v, na.rm=T)
  }
  v
}) %>% as.data.table()
```



# Create data container

``` {r}
sampleGroups <- as.factor(Sample_to_Group(sampleNames))
lipidClasses <- as.factor(Lipid_to_Class(colnames(lipidData)))

data <- list(dfs = list(lipids = lipidData,
                        proteins = proteinData),
             samples = data.table(names = sampleNames,
                                  group = sampleGroups),
             features = list(lipids = list(class = lipidClasses,
                                           ID = colnames(lipidData)),
                             proteins = list(ID = accessionIDs))
)



```


# Add sample metadata

## Read in metadata 

``` {r}
sampleMetaData <- paste0(wd, "Data/CASABLANCA all variables.xlsx") %>% 
  read_xlsx(sheet = 1) %>% 
  as.data.table()
```

## Homogenise sample names to agree with data container

First we extract the sample group and adjust `"No CAD"` to `"noCAD"`. We then combine it with the sample ID, removing the leading zeroes. Now the data container's samples and the metadata's samples should be in the same form. Given the intersection of the two dataframe's sample names is of length 70, we now know that all the samples are in the same format and exist across both datasets:

``` {r}
# Extract sample groups and replace "No CAD" with "noCAD" 
sampleGroups <- sampleMetaData$Group
sampleGroups[sampleGroups=="No CAD"] <- "noCAD"

# Create new sample names with sample group followed by ID. Then remove leading 0s from ID
sampleNames <- paste0(sampleGroups, "_", v(sampleMetaData$ID))
sampleNames <- gsub("(?<=_)0+", "", sampleNames, perl = TRUE)

# Check that intersection is of length 40
intersect(mixedsort(sampleNames), data$samples$names) %>% length()
```

Now we can reorder the metadata data frame:

``` {r}
sampleMetaData <- sampleMetaData[mixedorder(sampleNames)]
```

From this, we can extract some of the meta data variables (now ordered correctly) and add them to our data container. Additionally, to aid in future plotting, we will make a discretised form of the age variable:

``` {r}
data$samples$age <- as.numeric(sampleMetaData$age)
data$samples$ageDisc <- cut(sampleMetaData$age, 
                            breaks = c(40, 50, 60, 70, 80, 90))

data$samples$race <- as.factor(sampleMetaData$race)

data$samples$sex <- as.factor(sampleMetaData$sex)

data$samples$smoker <- as.factor(sampleMetaData$smoker)
```




# Save data container

``` {r}
data %>% saveRDS(paste0(wd, "RDS/Data_Objects/all_data.rds"))
```
